% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigWig2Cov.R
\name{bigWig2Cov}
\alias{bigWig2Cov}
\title{Conversion the file in bigWig format to list of gene coverage}
\usage{
bigWig2Cov(bw)
}
\arguments{
\item{bw}{annotation file}
}
\description{
Conversion the file in bigWig format to list of gene coverage
}
