% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadCoverage.R
\name{getReadCoverage}
\alias{getReadCoverage}
\title{Extract read coverage from bam files}
\usage{
getReadCoverage(
  bams,
  chr,
  start,
  end,
  strand = NA,
  scanBamFlags = list(),
  tagFilter = list()
)
}
\arguments{
\item{bams}{character vector with paths to bam-files (bai-files requires to be storage in the same directory)}

\item{chr}{contig name}

\item{start, end}{coordinates of region}

\item{strand}{strand, NA for unstranded (default)}

\item{scanBamFlags}{list of flags to filter reads, see GenomicAlignments::scanBamFlag}

\item{tagFilter}{list of tags value to filter by. Can be used to filter cells in single cell bams, for example tagFilter=list(CB=barcodes)}
}
\value{
list with three items: x (chr coordinates); cov - number of reads mapped to chr position; juncs - data.frame with introns
}
\description{
Extract read coverage from bam files
}
