% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecountCov.R
\name{getRecountCov}
\alias{getRecountCov}
\title{Recounting of coordinates}
\usage{
getRecountCov(sid, jxn, gene.grange)
}
\arguments{
\item{sid}{ID of sequence}

\item{jxn}{number of junction}

\item{gene.grange}{generange info}
}
\description{
Recounting of coordinates
}
