% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEnsGTF.R
\name{loadEnsGTF}
\alias{loadEnsGTF}
\title{Loading of genome annotation in ensemble format}
\usage{
loadEnsGTF(f, features = NULL)
}
\arguments{
\item{f}{path to file with ensemble annotation}

\item{features}{}
}
\description{
Loading of genome annotation in ensemble format
}
