% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadCov.R
\name{plotReadCov}
\alias{plotReadCov}
\title{Plots read coverage}
\usage{
plotReadCov(
  r,
  min.junc.cov = 0,
  min.junc.cov.f = 0,
  plot.junc.only.within = FALSE,
  ylim = NULL,
  xlim = range(r$x),
  reverse = FALSE,
  junc.col = "blue",
  junc.lwd = 3,
  bottom.mar = 0,
  ...
)
}
\arguments{
\item{r}{read coverage; output of \code{\link{getReadCoverage}}}

\item{min.junc.cov}{numeric, plots only junctions (introns) with coverage not less than \code{min.junc.cov}}

\item{min.junc.cov.f}{numeric, plots only junctions (introns) with coverage not less than \code{min.junc.cov.f} of maximal coverage in the region}

\item{plot.junc.only.within}{logical, plot only juction with both ends within the region, FALSE plots all junctions with at least one end within region. NA plot all junctions overlapping the region.}

\item{ylim, xlim}{see \code{\link{plot}}}

\item{reverse}{reverse x coordinates}

\item{junc.col}{colour for junction line. Individual color could be specified for each junction}

\item{junc.lwd}{line width for jucntion line}

\item{...}{other parameters for plot function}
}
\description{
Plots read coverage
}
