% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTranscripts.R
\name{plotTranscripts}
\alias{plotTranscripts}
\title{Plots transcripts for considered gene}
\usage{
plotTranscripts(
  a,
  ylim = c(0, length(unique(a$transcript_id))),
  xlim = c(ifelse(a$strand[1] == "+", min(a$start), max(a$stop)), ifelse(a$strand[1] ==
    "+", max(a$stop), min(a$start))),
  xlab = a$chr_id[1],
  new = TRUE,
  yspace = 0.8,
  exon.col = "black",
  cds.col = "black",
  text.cex = 0.7,
  ...
)
}
\arguments{
\item{a}{gene data from ensemble gtf annotation which transcripts are plotted}

\item{ylim, xlim}{see \code{\link{plot}}}

\item{xlab}{see \code{\link{plot}}}

\item{new}{logical: if to plot transcripts as new plot  TRUE|FALSE}

\item{yspace}{space in height}

\item{exon.col}{color of exon border/fill}

\item{cds.col}{color CDS line}

\item{text.cex}{see \code{\link{plot}}}

\item{...}{other parameters for plot function}

\item{reverse}{reverse x coordinates}
}
\description{
Plots transcripts for considered gene
}
